Fixed parsing full song names with special characters.
Fixed parsing song names when the name had an ' in it. strtok_s would only parse up to the ' and consider the next part the new p value breaking all following values.

Now we parse out the string and then feed that encapsulated value to getSongData.

removed explicit "|" and "^" in strotok_s calls we already have the value in a variable.